requirements.txt
meshtastic
python-dotenv 